--- src/app/app.component.scss
@@ -1,24 +1,0 @@
-@import '../injects';
-
-.app-component-outlet {
-
-	position:relative;
-	left:0px;
-	right:0px;
-
-	@include transition((right 0.5s,left .5s));
-
-	@media (min-width: $screen-desktop){
-		position: absolute;
-		left:400px;
-		right:0px;
-	}
-	@media (min-width: $screen-1366 + 1){
-		right:300px;
-	}
-	@media (min-width: $screen-xlg-desktop){
-		position: absolute;
-		left:450px;
-		right:350px;
-	}
-}
