--- src/app/app.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-root',
-  templateUrl: './app.component.html',
-  styleUrls: ['./app.component.scss'],
-})
-export class AppComponent implements OnInit {
-  constructor () {
-
-  }
-  ngOnInit (): void {
-    // this.communications.connect();
-  }
-}
