+++ src/app/core/login/login.component.spec.ts
@@ -1,11 +1,11 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-import { UserService } from '@app/services/user.service';
+import { UserService } from '@app/core/services/user.service';
 import { LoginComponent } from './login.component';
-import { CommunicateService } from '@app/services/communicate.service';
+import { CommunicateService } from '@app/core/services/communicate.service';
 import { StoreModule } from '@ngrx/store';
-import { RequestsService } from '@app/services/requests.service';
-import { PermissionsService } from '@app/services/permissions.service';
-import { MocksService } from '@app/services/mocks.service';
+import { RequestsService } from '@app/core/services/requests.service';
+import { PermissionsService } from '@app/core/services/permissions.service';
+import { MocksService } from '@app/core/services/mocks.service';
 import { RouterTestingModule } from '@angular/router/testing';
 import { RouterModule } from '@angular/router';
