+++ src/app/core/navigation/navigation.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
-import { UserService } from '@app/services/user.service';
-import { AuthGuard } from '@app/services/user.service';
+import { UserService } from '@app/core/services/user.service';
+import { AuthGuard } from '@app/core/services/user.service';
 @Component({
   selector: 'app-navigation',
   templateUrl: './navigation.component.html',
