+++ src/app/core/users/logout/logout.component.ts
@@ -1,6 +1,6 @@
 import { Router } from '@angular/router';
 import { Component, OnInit } from '@angular/core';
-import { UserService } from '@app/services/user.service';
+import { UserService } from '@app/core/services/user.service';
 @Component({
   selector: 'app-logout',
   templateUrl: './logout.component.html',
