+++ src/app/core/users/users.component.spec.ts
@@ -3,9 +3,9 @@ import { RouterTestingModule } from '@angular/router/testing';
 import { NO_ERRORS_SCHEMA } from '@angular/core';
 import { UsersComponent } from './users.component';
 import { RouterModule } from '@angular/router';
-import { RequestsService } from '@app/services/requests.service';
-import { PermissionsService } from '@app/services/permissions.service';
-import { MocksService } from '@app/services/mocks.service';
+import { RequestsService } from '@app/core/services/requests.service';
+import { PermissionsService } from '@app/core/services/permissions.service';
+import { MocksService } from '@app/core/services/mocks.service';
 
 describe('UsersComponent', () => {
   let component: UsersComponent;
