+++ src/app/iot/locations/locations.component.spec.ts
@@ -2,15 +2,15 @@ import { NO_ERRORS_SCHEMA } from '@angular/core';
 import { async, ComponentFixture, TestBed, inject } from '@angular/core/testing';
 import { StoreModule } from '@ngrx/store';
 import { PageHeaderComponent } from '../shared/page-header/page-header.component';
-import { RequestsService } from '@app/services/requests.service';
+import { RequestsService } from '@app/core/services/requests.service';
 import { LocationsComponent } from './locations.component';
 import { PlacesComponent } from './../index/places/places.component';
-import { CommunicateService } from '@app/services/communicate.service';
+import { CommunicateService } from '@app/core/services/communicate.service';
 import { appReducersGenerator } from '@app/core/app.reducers';;
 import { ActivatedRoute } from '@angular/router';
-import { PermissionsService } from '@app/services/permissions.service';
-import { MocksService } from '@app/services/mocks.service';
-import { ActionsService } from '@app/services/actions.service';
+import { PermissionsService } from '@app/core/services/permissions.service';
+import { MocksService } from '@app/core/services/mocks.service';
+import { ActionsService } from '@app/core/services/actions.service';
 import { Observable } from 'rxjs/Observable';
 import 'rxjs/add/observable/of';
