+++ src/app/iot/quick-status/user-widget/user-widget.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { UserService } from '@app/services/user.service';
+import { UserService } from '@app/core/services/user.service';
 @Component({
   selector: 'app-user-widget',
   templateUrl: './user-widget.component.html',
