+++ src/app/iot/shared/helper-box/helper-box.component.spec.ts
@@ -5,11 +5,11 @@ import { SelectInputComponent } from '../../forms/select-input/select-input.comp
 import { TextInputComponent } from '../../forms/text-input/text-input.component';
 import { appReducersGenerator } from '@app/core/app.reducers';
 
-import { CommunicateService } from '@app/services/communicate.service';
-import { RequestsService } from '@app/services/requests.service';
-import { PermissionsService } from '@app/services/permissions.service';
-import { MocksService } from '@app/services/mocks.service';
-import { ActionsService } from '@app/services/actions.service';
+import { CommunicateService } from '@app/core/services/communicate.service';
+import { RequestsService } from '@app/core/services/requests.service';
+import { PermissionsService } from '@app/core/services/permissions.service';
+import { MocksService } from '@app/core/services/mocks.service';
+import { ActionsService } from '@app/core/services/actions.service';
 
 
 describe('HelperBoxComponent', () => {
